pyproject.toml
[tool.poetry]
name = "ali-scraper"
version = "0.1.0"
description = "Scraper modulaire pour AliExpress"
authors = ["TonNom <ton.email@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
selenium = "^4.0"
webdriver-manager = "^4.0"
python-dotenv = "^0.19"
typer = "^0.4"
pandas = "^1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"